trim out interface bloat - it will be a separate module
