+++ .babelrc
@@ -1,4 +1,3 @@
 {
-  "presets": [["env", {"targets": {"node": 6}}], "react"],
-  "plugins": ["transform-runtime", "transform-object-rest-spread", "glamorous-displayname"]
+  "plugins": ["transform-object-rest-spread"]
 }
