--- interface/.storybook/preview-head.html
@@ -1,2 +1,0 @@
-<script src="https://use.typekit.net/qxd4ymv.js"></script>
-<script>try{Typekit.load({ async: true });}catch(e){}</script>
