--- interface/src/Presets.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-class Presets extends React.Component {
-  render () {
-    return (
-      'not implemented'
-    )
-  }
-}
-
-export default Presets
