--- interface/src/icons/AppIcon.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-
-const AppIcon = () =>
-  <svg width='43px' height='39px' viewBox='0 0 43 39'>
-    <g id='Page-1' stroke='none' strokeWidth='1' fill='none' fillRule='evenodd'>
-      <g id='Artboard' transform='translate(-43.000000, -37.000000)' fillRule='nonzero' fill='#7F8FA4'>
-        <g id='react' transform='translate(43.000000, 37.000000)'>
-          <path
-            d='M21.5009089,15.9351723 C23.5043787,15.9351723 25.1382764,17.56907 25.1382764,19.6114421 C25.1382764,21.5565583 23.5043787,23.2099072 21.5009089,23.2099072 C19.4974392,23.2099072 17.8635415,21.5565583 17.8635415,19.6114421 C17.8635415,17.56907 19.4974392,15.9351723 21.5009089,15.9351723 L21.5009089,15.9351723 Z M12.4950206,35.1723722 C13.7204439,35.9115164 16.4047043,34.7833489 19.4974392,31.8656745 C18.4859787,30.7180559 17.4939694,29.4731815 16.5603136,28.1699536 C14.9653183,28.0143443 13.3897741,27.7809304 11.8920346,27.4697118 C10.9000253,31.6322606 11.2695974,34.4915815 12.4950206,35.1723722 L12.4950206,35.1723722 Z M13.8760532,24.0074048 L13.3119695,23.0153955 C13.0980067,23.5794792 12.8840439,24.143563 12.7478857,24.6881955 C13.2730671,24.8049025 13.856602,24.9021583 14.4595881,24.9994141 C14.2650764,24.6687444 14.0705648,24.3380746 13.8760532,24.0074048 L13.8760532,24.0074048 Z M26.5971136,22.5291165 L28.1726577,19.6114421 L26.5971136,16.6937677 C26.0135787,15.662856 25.3911415,14.7486514 24.8270578,13.8344468 C23.776695,13.7760933 22.6679787,13.7760933 21.5009089,13.7760933 C20.3338392,13.7760933 19.2251229,13.7760933 18.1747601,13.8344468 C17.6106764,14.7486514 16.9882392,15.662856 16.4047043,16.6937677 L14.8291601,19.6114421 L16.4047043,22.5291165 C16.9882392,23.5600281 17.6106764,24.4742327 18.1747601,25.3884374 C19.2251229,25.4467909 20.3338392,25.4467909 21.5009089,25.4467909 C22.6679787,25.4467909 23.776695,25.4467909 24.8270578,25.3884374 C25.3911415,24.4742327 26.0135787,23.5600281 26.5971136,22.5291165 L26.5971136,22.5291165 Z M21.5009089,9.45793516 C21.1313368,9.88586074 20.7423136,10.3332375 20.3532903,10.8584189 C20.7423136,10.8584189 21.1118857,10.8584189 21.5009089,10.8584189 C21.8899322,10.8584189 22.2595043,10.8584189 22.6485275,10.8584189 C22.2595043,10.3332375 21.870481,9.88586074 21.5009089,9.45793516 L21.5009089,9.45793516 Z M21.5009089,29.764949 C21.870481,29.3370234 22.2595043,28.8896466 22.6485275,28.3644653 C22.2595043,28.3644653 21.8899322,28.3644653 21.5009089,28.3644653 C21.1118857,28.3644653 20.7423136,28.3644653 20.3532903,28.3644653 C20.7423136,28.8896466 21.1313368,29.3370234 21.5009089,29.764949 L21.5009089,29.764949 Z M30.4873461,4.05051194 C29.281374,3.31136776 26.5971136,4.4395352 23.5043787,7.3572096 C24.5158391,8.50482819 25.5078484,9.7497026 26.4415043,11.0529305 C28.0364996,11.2085398 29.6120438,11.4419538 31.1097833,11.7531724 C32.1017926,7.59062355 31.7322205,4.73130264 30.4873461,4.05051194 L30.4873461,4.05051194 Z M29.1257647,15.2154793 L29.6898484,16.2074886 C29.9038112,15.6434049 30.117774,15.0793212 30.2539321,14.5346886 C29.7287507,14.4179816 29.1452159,14.3207258 28.5422298,14.22347 C28.7367414,14.5541398 28.9312531,14.8848095 29.1257647,15.2154793 L29.1257647,15.2154793 Z M31.9461833,1.50240964 C34.8055042,3.1363073 35.1167228,7.43501425 33.9107507,12.4534142 C38.851346,13.9122514 42.4109088,16.3241956 42.4109088,19.6114421 C42.4109088,22.8986886 38.851346,25.3106327 33.9107507,26.7694699 C35.1167228,31.7878699 34.8055042,36.0865768 31.9461833,37.7204745 C29.1063135,39.3543722 25.2355322,37.4870605 21.5009089,33.9274978 C17.7662857,37.4870605 13.8955043,39.3543722 11.0361834,37.7204745 C8.19631369,36.0865768 7.88509508,31.7878699 9.09106717,26.7694699 C4.15047186,25.3106327 0.590909091,22.8986886 0.590909091,19.6114421 C0.590909091,16.3241956 4.15047186,13.9122514 9.09106717,12.4534142 C7.88509508,7.43501425 8.19631369,3.1363073 11.0361834,1.50240964 C13.8955043,-0.131488025 17.7662857,1.73582359 21.5009089,5.29538635 C25.2355322,1.73582359 29.1063135,-0.131488025 31.9461833,1.50240964 L31.9461833,1.50240964 Z M31.3820996,19.6114421 C32.0434391,21.0702793 32.626974,22.5291165 33.113253,24.0074048 C37.1979972,22.7819816 39.4932344,21.0313769 39.4932344,19.6114421 C39.4932344,18.1915072 37.1979972,16.4409026 33.113253,15.2154793 C32.626974,16.6937677 32.0434391,18.1526049 31.3820996,19.6114421 L31.3820996,19.6114421 Z M11.6197183,19.6114421 C10.9583788,18.1526049 10.3748439,16.6937677 9.88856484,15.2154793 C5.80382068,16.4409026 3.50858349,18.1915072 3.50858349,19.6114421 C3.50858349,21.0313769 5.80382068,22.7819816 9.88856484,24.0074048 C10.3748439,22.5291165 10.9583788,21.0702793 11.6197183,19.6114421 L11.6197183,19.6114421 Z M29.1257647,24.0074048 C28.9312531,24.3380746 28.7367414,24.6687444 28.5422298,24.9994141 C29.1452159,24.9021583 29.7287507,24.8049025 30.2539321,24.6881955 C30.117774,24.143563 29.9038112,23.5794792 29.6898484,23.0153955 L29.1257647,24.0074048 L29.1257647,24.0074048 Z M23.5043787,31.8656745 C26.5971136,34.7833489 29.281374,35.9115164 30.4873461,35.1723722 C31.7322205,34.4915815 32.1017926,31.6322606 31.1097833,27.4697118 C29.6120438,27.7809304 28.0364996,28.0143443 26.4415043,28.1699536 C25.5078484,29.4731815 24.5158391,30.7180559 23.5043787,31.8656745 L23.5043787,31.8656745 Z M13.8760532,15.2154793 C14.0705648,14.8848095 14.2650764,14.5541398 14.4595881,14.22347 C13.856602,14.3207258 13.2730671,14.4179816 12.7478857,14.5346886 C12.8840439,15.0793212 13.0980067,15.6434049 13.3119695,16.2074886 L13.8760532,15.2154793 L13.8760532,15.2154793 Z M19.4974392,7.3572096 C16.4047043,4.4395352 13.7204439,3.31136776 12.4950206,4.05051194 C11.2695974,4.73130264 10.9000253,7.59062355 11.8920346,11.7531724 C13.3897741,11.4419538 14.9653183,11.2085398 16.5603136,11.0529305 C17.4939694,9.7497026 18.4859787,8.50482819 19.4974392,7.3572096 Z'
-            id='Shape' />
-        </g>
-      </g>
-    </g>
-  </svg>
-
-export default AppIcon
