--- interface/src/icons/Replay.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Replay = () =>
-  <svg width='24' height='24' viewBox='0 0 24 24'>
-    <path
-      d='M12,5V1L7,6L12,11V7A6,6 0 0,1 18,13A6,6 0 0,1 12,19A6,6 0 0,1 6,13H4A8,8 0 0,0 12,21A8,8 0 0,0 20,13A8,8 0 0,0 12,5Z' />
-  </svg>
-
-export default Replay
