--- interface/src/todo
@@ -1,1 +1,0 @@
-load endpoint details for active route in endpoint details
