--- views/host.pug
@@ -1,12 +1,0 @@
-doctype html
-html
-  meta(name="viewport", content="width=device-width, initial-scale=1")
-  head
-    title Appstrap
-  body
-    if container.startsWith('#')
-      div(id= container.replace('#', ''))
-    else
-      div(class= container.replace('.', ''))
-
-    script(src= url)
