--- views/interface.pug
@@ -1,9 +1,0 @@
-doctype html
-html
-    head
-        script(src='https://use.typekit.net/qxd4ymv.js')
-        script.
-            try{Typekit.load({ async: true });}catch(e){}
-    body
-        div#react-host
-        script(src='/interface.js')
